+++ src/PWM_Generic_Debug.h
@@ -6,7 +6,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/RP2040_PWM
   Licensed under MIT license
 
-  Version: 1.4.1
+  Version: 1.5.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -23,6 +23,7 @@
   1.3.1   K Hoang      11/09/2022 Add minimal example `PWM_Basic`
   1.4.0   K Hoang      15/10/2022 Fix glitch when changing dutycycle. Adjust MIN_PWM_FREQUENCY/MAX_PWM_FREQUENCY dynamically
   1.4.1   K Hoang      21/01/2023 Add `PWM_StepperControl` example
+  1.5.0   K Hoang      24/01/2023 Add `PWM_manual` example and functions
  *****************************************************************************************************************************/
 
 #pragma once